+++ FLASK/hstack/uploadApi/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask
 from flask_restx import Api
 
-from . import extractMetadata
+from . import main
 
 
 def create_app():
@@ -11,10 +11,10 @@ def create_app():
     app.config.from_pyfile('config.py')
 
     # Namespace
-    api.add_namespace(extractMetadata.Upload, '/upload')
+    api.add_namespace(main.Upload, '/upload')
 
     return app
 
 if __name__ == '__main__':
     app = create_app()
-    app.run(host='0.0.0.0', port='8000', debug=True)
+    app.run()
