--- FLASK_TEST/test/__init__.py
@@ -1,19 +1,0 @@
-from flask import Flask
-from flask_restx import Api
-
-from test import extractMetadata
-
-def create_app():
-    app = Flask(__name__)
-    api = Api(app)
-
-    app.config.from_pyfile('config.py')
-
-    # Namespace
-    api.add_namespace(extractMetadata.Upload, '/')
-
-    return app
-
-if __name__ == '__main__':
-    app = create_app()
-    app.run()
