--- FLASK_TEST/test/models.py
@@ -1,6 +1,0 @@
-# coding: utf-8
-from flask_sqlalchemy import SQLAlchemy
-
-
-db = SQLAlchemy()
-
