--- FLASK_TEST/test/templates/success.html
@@ -1,1 +1,0 @@
-Upload Successed.
