--- FLASK_TEST/test/templates/upload.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>VMeta - File Upload</title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    </head>
-    <body>
-        <form
-            action="."
-            method="POST"
-            enctype="multipart/form-data">
-            <table id="uploadFileTable">
-                <tr>
-                    <td>
-                        <div class="tags">Title :
-                        </div>
-                    </td>
-                    <td class="uploadFileTableTD">
-                        <input
-                            type="text"
-                            name="fileTitle"
-                            class="inputText"
-                            placeholder="Enter a title">
-                    </td>
-                </tr>
-                <tr>
-                    <td class="uploadFileTableTD">
-                        <div class="tags">Presenter :
-                        </div>
-                    </td>
-                    <td class="uploadFileTableTD">
-                        <input type="text" name="presenter" class="inputText" placeholder="HSTACK">
-                    </td>
-                </tr>
-                <tr>
-                    <td class="uploadFileTableTD">
-                        <div class="tags">Video File :
-                        </div>
-                    </td>
-                    <td class="uploadFileTableTD">
-                        <input type="file" name="videoFile" id="btn" ;=";">
-                        <!-- <label id="fileBtn" for="btn"> 파일 선택 </label> -->
-                    </td>
-                </tr>
-                <tr>
-                    <td class="uploadFileTableTD"></td>
-                    <td id="uploadFileSubmit">
-                        <input id="inputSubmit" type="submit" value="Upload">
-                    </td>
-                </tr>
-            </table>
-        </form>
-    </body>
-</html>
